requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# HTTP & API clients
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.27.0

# Data processing
pydantic>=2.5.0
pandas>=2.1.0

# NLP & Text processing
nltk>=3.8.0
scikit-learn>=1.3.0

# Reddit API
praw>=7.7.0

# Twitter/X API (via tweepy or snscrape)
tweepy>=4.14.0

# Messaging
python-telegram-bot>=21.0
twilio>=9.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
asyncio>=3.4.3

# Scheduling
apscheduler>=3.10.0
